+++ backend/src/main/java/com/Helpers.java
@@ -1,4 +1,4 @@
-package test.java;
+package com;
 
 import java.util.ArrayList;
 import java.util.Date;
