+++ backend/src/main/java/com/datamodel.java
@@ -1,4 +1,4 @@
-package test.java;
+package com;
 import java.util.*;
